debugger2/requirements.txt
bidict==0.23.1
click==8.1.8
h11==0.14.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
pytest-asyncio==0.25.0
python-engineio==4.11.1
python-socketio==5.12.0
simple-websocket==1.1.0
uvicorn==0.34.0
wsproto==1.2.0
